only creator can edit and delete now

+++ app/src/main/java/cookbook/controller/CommentController.java
@@ -3,10 +3,10 @@ package cookbook.controller;
 import java.sql.SQLException;
 import java.text.SimpleDateFormat;
 
-import com.mysql.cj.xdevapi.Session;
-
 import cookbook.model.Comment;
 import cookbook.model.QueryMaker;
+import cookbook.model.Session;
+import cookbook.model.User;
 import javafx.event.ActionEvent;
 import javafx.fxml.FXML;
 import javafx.scene.control.Button;
@@ -32,38 +32,61 @@ public class CommentController {
   private Label myusername;
 
   private Comment comment; // Assume Comment class represents a comment object
-
-  private AnchorPane parentAnchorPane;
+  private User user;
 
   public void setData(Comment comment, AnchorPane parent) {
 
-    String username = myusername.getText();
+    String username = user.getUsername();
+    myusername.setText(username);
+
     this.comment = comment;
-    parentAnchorPane = parent;
+
     this.mycomment.setText(comment.getComment_text());
     SimpleDateFormat dateFormat = new SimpleDateFormat("yyyy-MM-dd");
+
     myDate.setText(dateFormat.format(comment.getDate()));
+    int userid = comment.getUser_id();
+
+    if (userid != Session.getCurrentUser().getUserId()) {
+      editButton.setDisable(true);
+      deleteButton.setDisable(true);
+    }
   }
 
   @FXML
   void editComment(ActionEvent event) throws SQLException {
     String updatedComment = mycomment.getText();
     int userid = comment.getUser_id();
-    comment.setComment_text(updatedComment);
-    QueryMaker qm = new QueryMaker();
-    qm.editComment(comment);
 
+    if (userid == Session.getCurrentUser().getUserId()) {
+
+      try {
+
+        comment.setComment_text(updatedComment);
+        QueryMaker qm = new QueryMaker();
+        qm.editComment(comment);
+      } catch (SQLException e) {
+
+      }
+
+    } else {
+      System.out.println("only the creator of the comment can edit it.");
+    }
   }
 
   @FXML
   void removeComment(ActionEvent event) {
-    try {
-      QueryMaker qm = new QueryMaker();
-      qm.deleteComment(comment);
-    } catch (SQLException e) {
-      System.out.println("Error: " + e.getMessage());
+    int userid = comment.getUser_id();
+    if (userid == Session.getCurrentUser().getUserId()) {
+      try {
+        QueryMaker qm = new QueryMaker();
+        qm.deleteComment(comment);
+      } catch (SQLException e) {
+        System.out.println("Error: " + e.getMessage());
+      }
+    } else {
+      System.out.println("only the creator of the comment can delete it");
     }
-
   }
 
 }
